Move transaction based logic to separate module
