+++ mythril/laser/ethereum/svm.py
@@ -1,15 +1,13 @@
-from z3 import BitVec
 import logging
-from mythril.disassembler.disassembly import Disassembly
-from mythril.laser.ethereum.state import GlobalState, Environment, CalldataType, Account, WorldState
-from mythril.laser.ethereum.transaction import MessageCallTransaction, TransactionStartSignal, TransactionEndSignal, \
+from mythril.laser.ethereum.state import WorldState
+from mythril.laser.ethereum.transaction import TransactionStartSignal, TransactionEndSignal, \
     ContractCreationTransaction
 from mythril.laser.ethereum.instructions import Instruction
 from mythril.laser.ethereum.cfg import NodeFlags, Node, Edge, JumpType
 from mythril.laser.ethereum.strategy.basic import DepthFirstSearchStrategy
 from datetime import datetime, timedelta
-from functools import reduce
 from copy import copy
+from mythril.laser.ethereum.transaction import execute_contract_creation, execute_message_call
 
 TT256 = 2 ** 256
 TT256M1 = 2 ** 256 - 1
@@ -67,15 +65,15 @@ class LaserEVM:
 
         if main_address:
             logging.info("Starting message call transaction to {}".format(main_address))
-            self.execute_message_call(main_address)
+            execute_message_call(self, main_address)
         elif creation_code:
             logging.info("Starting contract creation transaction")
-            created_account = self.execute_contract_creation(creation_code)
+            created_account = execute_contract_creation(self, creation_code)
             logging.info("Finished contract creation, found {} open states".format(len(self.open_states)))
 
             self.time = datetime.now()
             logging.info("Starting message call transaction")
-            self.execute_message_call(created_account.address)
+            execute_message_call(self, created_account.address)
 
         logging.info("Finished symbolic execution")
         logging.info("%d nodes, %d edges, %d total states", len(self.nodes), len(self.edges), self.total_states)
@@ -239,64 +237,3 @@ class LaserEVM:
             return function
 
         return hook_decorator
-
-    def execute_message_call(self, callee_address):
-        """ Executes a message call transaction from all open states """
-        open_states = self.open_states[:]
-        del self.open_states[:]
-
-        for open_world_state in open_states:
-
-            transaction = MessageCallTransaction(
-                open_world_state,
-                open_world_state[callee_address],
-                BitVec("caller", 256),
-                [],
-                BitVec("gas_price", 256),
-                BitVec("call_value", 256),
-                BitVec("origin", 256),
-                CalldataType.SYMBOLIC,
-            )
-            self._setup_global_state_for_execution(transaction)
-
-        self.exec()
-
-    def execute_contract_creation(self, contract_initialization_code):
-        """ Executes a contract creation transaction from all open states"""
-        open_states = self.open_states[:]
-        del self.open_states[:]
-
-        new_account = self.world_state.create_account(0, concrete_storage=True)
-
-        for open_world_state in open_states:
-            transaction = ContractCreationTransaction(
-                open_world_state,
-                BitVec("caller", 256),
-                new_account,
-                Disassembly(contract_initialization_code),
-                [],
-                BitVec("gas_price", 256),
-                BitVec("call_value", 256),
-                BitVec("origin", 256),
-                CalldataType.SYMBOLIC
-            )
-
-            self._setup_global_state_for_execution(transaction)
-        self.exec(True)
-
-        return new_account
-
-    def _setup_global_state_for_execution(self, transaction):
-        """ Sets up global state and cfg for a transactions execution"""
-        global_state = transaction.initial_global_state()
-        global_state.transaction_stack.append((transaction, None))
-
-        new_node = Node(global_state.environment.active_account.contract_name)
-
-        self.nodes[new_node.uid] = new_node
-        if transaction.world_state.node:
-            self.edges.append(Edge(transaction.world_state.node.uid, new_node.uid, edge_type=JumpType.Transaction,
-                                   condition=None))
-        global_state.node = new_node
-        new_node.states.append(global_state)
-        self.work_list.append(global_state)
